backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
apscheduler
email-validator
notion-client
aiofiles
